requirements.txt
apache-airflow==2.7.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
imbalanced-learn==0.11.0